src/Jhipster/ClientApp/src/app/entities/job-history/delete/job-history-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { JobHistoryService } from '../service/job-history.service';

import { JobHistoryDeleteDialogComponent } from './job-history-delete-dialog.component';

describe('JobHistory Management Delete Component', () => {
  let comp: JobHistoryDeleteDialogComponent;
  let fixture: ComponentFixture<JobHistoryDeleteDialogComponent>;
  let service: JobHistoryService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, JobHistoryDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(JobHistoryDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(JobHistoryDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(JobHistoryService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
